fix: 修复 CI 构建错误（spec TOC 格式 + Rust macOS 编译）
- callwarden.spec: 按名称提取入口脚本（PyInstaller 6.x scripts 混入 rthook，固定下标取到 rthook tuple 导致 _normalize_toc ValueError）

- server.rs: socket_mode as u32 显式转换（macOS mode_t=u16，Permissions::from_mode 期望 u32）

+++ rust_ext/src/daemon/server.rs
@@ -185,9 +185,10 @@ where
 
     let listener = UnixListener::bind(&config.socket_path)?;
     // 设置 socket 文件权限（0o660：owner + group 可读写）
+    // macOS 上 libc::mode_t 是 u16，但 Permissions::from_mode 期望 u32，需要显式转换
     std::fs::set_permissions(
         &config.socket_path,
-        std::fs::Permissions::from_mode(config.socket_mode),
+        std::fs::Permissions::from_mode(config.socket_mode as u32),
     )?;
 
     // P0-3 v2 修复：socket chown 到指定组（多用户 UDS 访问）—— fail-closed
